P047 feat: add statuserror to keep track of status code

+++ Projects/07-nodejs/047-mongoose/controllers/products.js
@@ -1,16 +1,24 @@
 const Product = require('../models/product');
+const mongoose = require('mongoose');
+const StatusError = require('../helpers/StatusError');
 
 const getProducts = async (limit, skip, status) => {
-    if (limit && skip && status)
-        return await Product.find({ 'status': status }).limit(limit).skip(skip);
-    if (limit && skip)
-        return await Product.find().limit(limit).skip(skip);
+    let products;
+    if (status)
+        products = await Product.find({ 'status': status });
+    else
+        products = await Product.find();
+    if (limit)
+        products = await products.limit(limit);
+    if (skip)
+        products = await products.skip(skip);
+    return await products;
 }
 
 const createProduct = async (data, userId) => {
     const existingName = await Product.findOne({ 'name': data.name });
     if (existingName)
-        throw new Error("A product with this name already exists");
+        throw new StatusError(409, "A product with this name already exists");
     const newProduct = {
         "name": data.name,
         "quantity": data.quantity,
@@ -22,16 +30,18 @@ const createProduct = async (data, userId) => {
 };
 
 const editProduct = async (id, body, userId) => {
+    if (!mongoose.Types.ObjectId.isValid(id))
+        throw new StatusError(400, 'Invalid Product ID');
     const product = await Product.findById(id);
     if (!product)
-        throw new Error("This product does not exist");
+        throw new StatusError(404, "This product does not exist");
     if (product.owner.toString() !== userId) {
-        throw new Error("You are not authorized to edit this product");
+        throw new StatusError(401, "Unauthorized to edit this product");
     }
     if (body.name) {
         const existingName = Product.findOne({ 'name': body.name });
         if (existingName)
-            throw new Error("A product with this name already exists");
+            throw new StatusError(409, "A product with this name exists");
         product.name = body.name;
     }
     product.save();
@@ -40,32 +50,36 @@ const editProduct = async (id, body, userId) => {
 }
 
 const editProductStock = async (id, operation, quantity, userId) => {
+    if (!mongoose.Types.ObjectId.isValid(id))
+        throw new StatusError(400, "Invalid Product ID");
     const existingProduct = await Product.findById(id);
     if (!existingProduct)
-        throw new Error('This product does not exist');
+        throw new StatusError(404, "This product does not exist");
     if (existingProduct.owner.toString() !== userId)
-        throw new Error("You are not authorized to edit this product");
+        throw new StatusError(401, "Unauthorized to edit this product");
     if (operation === "restock" && existingProduct.quantity + quantity > 100)
-        throw new Error("Stock cannot exceed 100");
+        throw new StatusError(400, "Stock cannot exceed 100");
     if (operation === "destock" && existingProduct.quantity - quantity < 0)
-        throw new Error("Stock cannot be negative");
+        throw new StatusError(400, "Stock cannot be negative");
     if (operation === "restock") {
         existingProduct.quantity += quantity;
     } else if (operation === "destock") {
         existingProduct.quantity -= quantity;
     } else {
-        throw new Error("Invalid operation");
+        throw new StatusError(400, "Invalid operation");
     }
     await existingProduct.save();
     return existingProduct;
 }
 
 const deleteProduct = async (id, userId) => {
+    if (!mongoose.Types.ObjectId.isValid(id))
+        throw new StatusError(400, 'Invalid Product ID');
     const existingProduct = await Product.findById(id);
     if (!existingProduct)
-        throw new Error('This product does not exist');
+        throw new StatusError(404, 'This product does not exist');
     if (existingProduct.owner.toString() !== userId)
-        throw new Error("You are not authorized to delete this product");
+        throw new StatusError(401, "Unauthorized to delete this product");
     await Product.deleteOne({ '_id': id });
     return existingProduct;
 }
