+++ Projects/07-nodejs/047-mongoose/controllers/users.js
@@ -1,5 +1,7 @@
 const Product = require('../models/product');
 const User = require('../models/user');
+const mongoose = require('mongoose');
+const StatusError = require('../helpers/StatusError');
 
 const getUsers = async () => {
     const usersFirstName = await User.find().select('firstName');
@@ -7,14 +9,19 @@ const getUsers = async () => {
 }
 
 const getProductsOfUser = async (userId) => {
+    if (!mongoose.Types.ObjectId.isValid(userId))
+        throw new StatusError(400, 'Invalid User ID');
     const userExists = await User.findById(userId);
     if (!userExists)
-        throw new Error('This user does not exist');
+        throw new StatusError(404, 'This user does not exist');
     const productsOfUser = await Product.find({ 'owner': userId });
     return productsOfUser;
 }
 
 const createUser = async (data) => {
+    const existingUser = await User.findOne({ 'username': data.username });
+    if (existingUser)
+        throw new StatusError(409, 'This username is already taken');
     const user = await User.create({
         "firstName": data.firstName,
         "lastName": data.lastName,
@@ -25,9 +32,11 @@ const createUser = async (data) => {
 }
 
 const editUser = async (id, data) => {
+    if (!mongoose.Types.ObjectId.isValid(id))
+        throw new StatusError(400, 'Invalid User ID');
     const existingUser = await User.findById(id);
     if (!existingUser)
-        throw new Error('This user does not exist');
+        throw new StatusError(404, 'This user does not exist');
     if (data.firstName) {
         existingUser.firstName = data.firstName;
     }
@@ -45,9 +54,11 @@ const editUser = async (id, data) => {
 }
 
 const deleteUser = async (id) => {
+    if (!mongoose.Types.ObjectId.isValid(id))
+        throw new StatusError(400, 'Invalid User ID');
     const existingUser = await User.findById(id);
     if (!existingUser)
-        throw new Error('This user does not exist');
+        throw new StatusError(404, 'This user does not exist');
     await User.deleteOne({ '_id': id });
     return existingUser;
 }
